Move dash_email to utils.email and remove unused import
