+++ dash/orgs/models.py
@@ -3,21 +3,21 @@ from __future__ import unicode_literals
 import json
 import random
 
-from dash.dash_email import send_dash_email
-from dash.utils import datetime_to_ms
 from datetime import datetime
 from django.conf import settings
 from django.contrib.auth.models import User, Group
 from django.core.cache import cache
 from django.db import models
 from django.utils import timezone
-from django.utils.translation import ugettext_lazy as _
 from django.utils.encoding import force_text, python_2_unicode_compatible
+from django.utils.translation import ugettext_lazy as _
 from smartmin.models import SmartModel
 from temba_client.v1 import TembaClient as TembaClient1
 from temba_client.v2 import TembaClient as TembaClient2
 from timezone_field import TimeZoneField
 
+from dash.utils import datetime_to_ms
+from dash.utils.email import send_dash_email
 
 STATE = 1
 DISTRICT = 2
